fix spec
