+++ app/packs/src/decidim/decidim_awesome/admin/custom_fields_builder.js
@@ -1,5 +1,7 @@
 import "formBuilder/dist/form-builder.min.js";
 import "src/decidim/decidim_awesome/forms/rich_text_plugin"
+import getCustomFieldsLang from "src/decidim/decidim_awesome/forms/custom_fields_langs";
+
 // formBuilder uses jquery-ui-sortable which is a very dirty npm package with no neat source code available, and causes problems with the webpacker configuration of Decidim.
 // For the moment, we'll remove the sortable functionality with a dummy jQuery plugin until we find another sortable plugin (or keep it disabled for good)
 jQuery.fn.sortable = () => {}
@@ -10,6 +12,7 @@ $(() => {
   $(".awesome-edit-config .proposal_custom_fields_editor").each((_idx, el) => {
     const key = $(el).closest(".proposal_custom_fields_container").data("key");
     const configVar = $(el).closest(".proposal_custom_fields_container").data("var");
+    const lang = getCustomFieldsLang(window.DecidimAwesome.currentLocale);
     // DOCS: https://formbuilder.online/docs
     window.CustomFieldsBuilders.push({
       el: el,
@@ -17,8 +20,8 @@ $(() => {
       var: configVar,
       config: {
         i18n: {
-          locale: "en-US",
-          location: "https://cdn.jsdelivr.net/npm/formbuilder-languages@1.1.0/"
+          locale: lang,
+          location: window.DecidimAwesome.formBuilderLangsLocation
         },
         formData: $(`input[name="config[${configVar}][${key}]"]`).val(),
         disableFields: ["button", "file"],
@@ -76,6 +79,7 @@ $(() => {
 
   $("form.awesome-edit-config").on("submit", () => {
     window.CustomFieldsBuilders.forEach((builder) => {
+      console.log("builder", builder);
       $(`input[name="config[${builder.var}][${builder.key}]"]`).val(builder.instance.actions.getData("json"));
     });
   });
