+++ package.json
@@ -45,7 +45,7 @@
   "dependencies": {
     "codemirror": "^5.65.16",
     "form-storage": "^1.3.5",
-    "formBuilder": "^3.19.12",
+    "formBuilder": "^3.19.13",
     "inline-attacher": "^0.0.7",
     "jsrender": "^1.0.9",
     "leaflet.featuregroup.subgroup": "^1.0.2",
